Refactor and parse through Editor component
