+++ client/src/utils/debounce.js
@@ -1,0 +1,15 @@
+/*
+    Debouncing is a programming pattern or a technique to restrict the calling of a time-consuming function frequently, by delaying the execution of the function until a specified time to avoid unnecessary CPU cycles, and API calls and improve performance.
+    */
+export function debounce(duration, fn) {
+  let timeout;
+  return function (...args) {
+    if (timeout) {
+      clearTimeout(timeout);
+    }
+    timeout = setTimeout(() => {
+      timeout = null;
+      fn(args);
+    }, duration);
+  };
+}
